precision

+++ temperature_conv/src/lib.rs
@@ -3,7 +3,7 @@ pub fn fahrenheit_to_celsius(f: f64) -> f64 {
 }
 
 pub fn celsius_to_fahrenheit(c: f64) -> f64 {
-    (c*9 as f64/5 as f64)+32 as f64
+    (c*9.0/5.0)+32.0
 }
 #[cfg(test)]
 mod tests {
